+++ apps/wallet/data/tokens/src/main/java/com/tonapps/wallet/data/token/entities/TokenRateEntity.kt
@@ -4,7 +4,7 @@ import com.tonapps.wallet.data.core.WalletCurrency
 
 data class TokenRateEntity(
     val currency: WalletCurrency,
-    val fiat: Float,
+    val fiat: Double,
     val rate: Float,
     val rateDiff24h: String
 )
